backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.49
pydantic==2.12.5
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.1
anthropic==0.26.0
